Implement ContentLimit::BytesSoftWrap
Introduce a "soft wrap" content limit mode.

This mode acts like a soft wrap in text editors, but given a generic separator character allows soft wrapping new-lines as well.

For example this sentance:

The quick brown fox jumps over the lazy dog
Would break into the following parts given bytes=10 and separator=' ':

The quick 
brown fox 
jumps over 
the lazy dog
Note that the lazy dog is 12 bytes - which exceeds the soft limit of 10.

This is particularly useful when wrapping FileRotate with a BufWriter and writing jsons as lines, so as long as the FileRotate object is dropped - all rotated files are guaranteed to have complete (soft wrapped) writes, i.e valid jsons.

Note: When FileRotate is initialized with a pre-existing non-empty file that doesn't end with a separator - it rotates it immediately instead of handling previous partial writes. This can happen if a previous write session terminated abruptly without proper flushing.

Signed-off-by: Netanel Rosenthal <netanel.rosenthal@wiz.io>